% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svyEDER.R
\name{errEDER}
\alias{errEDER}
\title{Precisiones estadísticas, EDER 2011}
\usage{
errEDER(u, data = parent.frame())
}
\arguments{
\item{u}{Nombre de la variable.}

\item{data}{Base de datos.}
}
\value{
Matriz de resultados.
}
\description{
Pruebas de significancia estadística de los indicadores de la Encuesta Demográfica Retrospectiva (EDER) 2011
}
\details{
Para cada categoria, la funcion genera: la estimacion del indicador (Estim); error estandar (ErrStd); coeficiente de variación (CV); limite inferior (LInf(90%));limite superior (LSup(90%)); evaluacion significancia (Signif?)
}
\note{
Cualquier sugerencia es bienvenida
}
\author{
Julio C.
}

