% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svyINEGI.R
\name{svyINEGI.ENOE}
\alias{svyINEGI.ENOE}
\title{Precisiones estadísticas, ENOE}
\usage{
svyINEGI.ENOE(u, data = parent.frame())
}
\arguments{
\item{u}{Nombre de la variable.}

\item{data}{Base de datos.}
}
\value{
Matriz de resultados.
}
\description{
Pruebas de significancia estadística de los indicadores de la Encuesta Nacional de Ocupación y Empleo (ENOE)
}
\details{
Para cada categoria, la funcion genera: la estimacion del indicador (Estim); error estandar (ErrStd); coeficiente de variación (CV); limite inferior (LInf(90%));limite superior (LSup(90%)); evaluacion significancia (Signif?)
}
\note{
Cualquier sugerencia es bienvenida
}
\author{
Julio C.
}

