% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svyEDER.R
\name{errEDER}
\alias{errEDER}
\title{Precisiones estadisticas EDER.}
\usage{
errEDER(u, data = parent.frame())
}
\arguments{
\item{u}{Nombre de la variable.}

\item{data}{Base de datos.}
}
\value{
Matriz de resultados.
}
\description{
Precisiones estadisticas EDER.
}

