% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svyINEGI.R
\name{svyINEGI.ENOE}
\alias{svyINEGI.ENOE}
\title{Precisiones estadísticas.}
\usage{
svyINEGI.ENOE(u, data = parent.frame())
}
\arguments{
\item{u}{Nombre de la variable.}

\item{data}{Base de datos.}
}
\value{
Matriz de resultados.
}
\description{
Precisiones estadísticas.
}
\examples{
library(svyINEGI)
data(SDEMT115)
svyINEGI.ENOE("CLASE2",SDEMT115)
}

